back-end/requirements.txt
Django==5.1.7
python-chess==1.999
python-dotenv==1.0.1
gunicorn==23.0.0
django-allauth==65.1.0
dj-rest-auth==6.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.35.0
requests==2.32.3
cryptography==44.0.1
